* Restructured tests into their own source path... Gradle-style
