--- src/index.ts
@@ -1,3 +1,0 @@
-export { CHANGELOG } from './changelog'
-
-export * from './util'
