+++ src/main/api/badge-data.ts
