+++ src/main/api/badge-partial-data.ts
