+++ src/main/api/badge-partial-type.ts
