+++ src/main/api/badge-type.ts
