+++ src/main/api/game-map-data.ts
