+++ src/main/api/vidiot-map-point-of-interest-data.ts
