+++ src/main/changelog.ts
@@ -8,9 +8,8 @@ export const CHANGELOG: Change[] = [
       + '* Replaced redundant interfaces with their concrete equivalents.\n'
       + '* Replaced enums with extensible union types; Server groups with new badge types, enhancement types, etc. can now extend them locally.\n'
       + '* Removed the `serverGroup` property from entities to simplify the object tree; Server group context will need to be managed separately.\n'
-      + '* Package divided into two core modules: `api` - for integrators creating content packages, and `db` for integrators who want to consume those packages.\n'
       + '* Standardized pluralization of some field names (name, icon).\n'
-      + '* Combined `settitle` ids into a single field.\n'
+      + '* Combined `settitle` ids into a single tuple field.\n'
       + '* Change from GNU to The Unlicense.\n'
       + '* Removed dependency on lodash. There are now no third-party runtime dependencies.\n'
       + '* Moved from webpack to rollup for packaging.\n'
