+++ src/main/db/archetype.ts
