+++ src/main/db/badge-partial.ts
