+++ src/main/db/badge.ts
