+++ src/main/db/coh-content-database.ts
