+++ src/main/db/vidiot-map-point-of-interest.ts
